fix: perbaikan sebelum deploy (env, gitignore, migration, auth, nextjs cleanup)

+++ database/migrations/2026_05_20_011251_add_payment_columns_to_transactions_table.php
@@ -11,22 +11,22 @@ return new class extends Migration
         Schema::table('transactions', function (Blueprint $table) {
             // Kita tambahkan kolom-kolom yang belum ada (gunakan jika belum didefinisikan)
             if (!Schema::hasColumn('transactions', 'property_price')) {
-                $table->bigInteger('property_price')->default(0)->after('property_id');
+                $table->decimal('property_price', 15, 2)->default(0)->after('property_id');
             }
             if (!Schema::hasColumn('transactions', 'tax_amount')) {
-                $table->bigInteger('tax_amount')->default(0)->after('property_price');
+                $table->decimal('tax_amount', 15, 2)->default(0)->after('property_price');
             }
             if (!Schema::hasColumn('transactions', 'gross_amount')) {
-                $table->bigInteger('gross_amount')->default(0)->after('tax_amount');
+                $table->decimal('gross_amount', 15, 2)->default(0)->after('tax_amount');
             }
             if (!Schema::hasColumn('transactions', 'total_payable')) {
-                $table->bigInteger('total_payable')->default(0)->after('gross_amount');
+                $table->decimal('total_payable', 15, 2)->default(0)->after('gross_amount');
             }
             if (!Schema::hasColumn('transactions', 'amount_paid')) {
-                $table->bigInteger('amount_paid')->default(0)->after('total_payable');
+                $table->decimal('amount_paid', 15, 2)->default(0)->after('total_payable');
             }
             if (!Schema::hasColumn('transactions', 'admin_fee')) {
-                $table->integer('admin_fee')->default(0)->after('amount_paid');
+                $table->decimal('admin_fee', 15, 2)->default(0)->after('amount_paid');
             }
             if (!Schema::hasColumn('transactions', 'payment_type')) {
                 $table->string('payment_type')->default('cash')->after('payment_status');
